requirements.txt
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
fonttools==4.39.0
idna==3.4
ipython==8.10.0
jedi==0.18.2
kiwisolver==1.4.4
matplotlib==3.7.1
matplotlib-inline==0.1.6
mysql-connector-python==8.0.32
numpy==1.24.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.36
protobuf==3.20.3
pure-eval==0.2.2
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
stack-data==0.6.2
traitlets==5.9.0
urllib3==1.26.14
wcwidth==0.2.6
